feat: added gobilda headlight code to teleop

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/shooter/ShooterSubsystem.java
@@ -44,7 +44,7 @@ public class ShooterSubsystem {
         hw.shooter.setVelocity(0);
     }
 
-    public void setMaxRPM(int maxRPM){
+    public void setMaxRPM(double maxRPM){
         targetRPM = maxRPM;
     }
